Optimise away save() calls by calling it once at the end if needed.
Change the Settings API to not save the settings in every setter, but
to return true/false depending on if the value was changed. Then save
the settings only once, if needed, at the end of any public function
that used to save the settings.

-Add unit tests to verify that restore() and accessors do not
invoke save().
-Add unit test to verify that the current setter in the Speller
immediately saves the new value to the global Sonnet QSettings as it
used to before the change.

REVIEW: 123341