Reduce temporary allocations.
Access setting's default language directly without constructing
temporary Speller objects, which is extremely costly with the
internal dictionary loading.

Also, don't create the speller object and then set the language
afterwards, pass it directly to the ctor. This removes a potential
pessimization when the language is not equal to the default language.