Disable automatic disabling of automated checking by default.
Only check for disabling of automated checking if the error limits are set
to > 0, set them to 0 by default.

The old reason for it was apparently that opening e. g. source files in
Kate was slow, but Kate handles spell checking intelligently now, so it
doesn't seem to be a problem. Applications that have a problem with it
can turn it back on with the Setting API.