Don't cause circular linking on Windows
When building plugins, don't arbitrarily rename the output files
(as this will occasionally result in circular dependencies).

In this review[1] from three and a half years ago, the Sonnet plugins
were moved into a subdirectory, which was great. They were, however,
also renamed at the same time, which resulted in some (and
specifically the aspell plugin) being named the same as the library
they should be linked against. End result: Spell checking has not
worked on Windows for three and a half years.

[1] https://git.reviewboard.kde.org/r/118084/

Test plan:
With these properties set, the plugins fail to load on Windows
due to circular dependencies
Without them (that is, with this patch), the plugins load

Differential Revision: https://phabricator.kde.org/D9018