Don't impose using the default client
With this change, Sonnet's default client concept and the actual
behaviour are in better correspondence. The backend defined by the
defaultClient key in the Sonnet settings file will now be used only
if it supports the requested language. If not, Sonnet will fall back
to whatever backend is most reliable for that language.

Differential Revision: https://phabricator.kde.org/D11183