Sonnet: setLanguage should schedule a rehighlight if highlight is enabled
Summary:
... even if automatic enabling/disabling is off, as is the case in KMail.

This will allow removing a direct call to rehighlight() in kpimtextedit
which leads to the full highlighting being done twice (very slow on very
large emails).

Test Plan:
Replying to a mail in kmail and watching how many times
slotRehighlight is called. Well, this commit alone makes it worse,
but after a fix in kpimtextedit, it's down from 2 to 1.

Reviewers: sandsmark, mlaurent

Reviewed By: mlaurent

Subscribers: kde-frameworks-devel, #kde_pim

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D14447