Handle the case if createSpeller is passed an unavailable language
Summary:
In Loader::createSpeller, retun early if lang isn't one of the available
languages(); otherwise a dud entry will be created for it in
dictionarycombobox.

Emit a signal when loading a dictionary fails, so that Ui parts can
display an appropriate error message informing the user about the issue,
instead of failing silently.

This is a first step to fix bug 325541.

Reviewers: sandsmark, cullmann, dhaumann

Reviewed By: cullmann

Subscribers: dhaumann, cullmann, loh.tar, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D18907