Allow overriding to disable auto language detection
Summary:
Currently, if language detection is enabled at Sonnet-wide level level, there is no way to disable it in a specific context.

In Lokalize for instance, the language of the source and target strings are already known by the application.

In that context, the language detection feature creates a lot of uncertainty in terms of spellchecking, as it may change quickly from one language to another or from one dictionnary to another within the same language. A user should not have to disable Sonnet-wide language detection to work around this, as it can be very useful elsewhere.

https://bugs.kde.org/show_bug.cgi?id=394347
BUG:394347

Reviewers: #frameworks, cullmann, mlaurent, mludwig

Subscribers: aacid, mludwig, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D26342