hunspell: actually use non-alias symlinks to dictionaries
Previously, if a dictionary was a symlink, but _not_ an alias,
it would just be silently ignored. This breaks loudly on systems
like NixOS, where dictionaries are basically always symlinks,
but is also just generally confusing and feels unintended.

This changes the logic to just add those as normal dictionaries.

Co-authored-by: jys1670 <jys1670@hotmail.com>
Signed-off-by: Ilya Katsnelson <me@0upti.me>