[hunspell] Don't treat symlinks to dict files specially
Currently the code assumes that when a dict file is a symlink that it's
an alias for another language. That breaks when it's a symlink to a file
somewhere completely different.

For example in Flatpak /usr/share/hunspell/de_DE.aff is a symlink to
/usr/share/runtime/locale/de/share/de/hunspell/de_DE_frami.aff

This breaks loading the file because it never gets added to m_languagePaths

That said, we don't treat languages and aliases particularly different,
so we can get rid of the alias list and treat every file as its own language.

This both fixes the bug and simplifies the code.

BUG: 507931