Ensure parent is alive when removing eventfilters
Scenario 1: The QTextEdit/QPlainTextEdit parent of the
SpellCheckDecorator is destroyed. In that case the
q pointer of SpellCheckDecoratorPrivate, m_textEdit,
and m_plainTextEdit are already invalid when the
SpellCheckDecoratorPrivate destructor are called. This
leads to memory corruption when trying to remove eventfilters
on the already invalid m_textEdit and m_plainTextEdit. Below
are output from a small sample program with QDebug statements
in all destructors.

  TextEdit destructor
  SpellCheckDecorator destructor
  SpellCheckDecoratorPrivate destructor
  corrupted double-linked list
  Aborted                    (core dumped) ./scenario1

Scenario 2: The SpellCheckDecorator is destroyed while the parent
QTextEdit/QPlainTextEdit lives on. In this case it probably makes
sense to remove the event filters, but this should be done while
the q pointer is still valid.

Without this patch Trojitá crashes when exiting the mail composer when
using the Sonnet spellchecker plugin.

BUG: 492444